Create README.md
Hi, I'm Tanzeela konian. I'm Looking for an opportunity to utilize my acquired skills and training to help the company grow and also I want to use my knowledge to gain practical exposure and understand the inner-workings of the company. I personally like to learn things continuously, so I'm looking forward to it.